Release 1.1.0
Welcome to Hardcore Score version 1.1.0

This new version brings Leaderboards to Hardcore Score.  In this first roll out it is limited to Guilds, Parties and Raids.  We will open it up to everyone in the next version (soon).  Since it is hard sometime fully test these leaderboards you may encounter some bugs.  Please be patient as we iron those out.  If you encounter a problem it would be appreciated if you could report it on our Discord ( https://discord.gg/j92hrVZU2Q ).  This is a big step forward for Hardcore Score.  It will set up what we need to bring more fun in your Hardcore journey.  Please enjoy and let us know what you think.

    1. New Scoreboard/Leaderboard (for guilds only)
    2. New Character Journey information
    3. New view of Milestones
    4. New view of Achievements
    5. Level 60 bug fixes
    6. Overall just tons and tons of updates to the code.
    7. Ability to reset leaderboard by using /hcsresetl
    8. Updated for version 1.1.5 of Classic Era
    9. Updated for version 3.4.3 of Classic Era (WOTLK)
  10. New options for sharing your Hardcore Score.